+++ python/src/scipp/plotting/resampling_model.py
@@ -134,7 +134,7 @@ class ResamplingBinnedModel(ResamplingModel):
             if len(var.dims) == 0:
                 continue
             dim = var.dims[-1]
-            if name not in self._array.data.bins.data.meta:
+            if name not in self._array.data.bins.meta:
                 self._array.meta[name] = to_bin_centers(var, dim)
 
     def _resample(self, array):
@@ -143,7 +143,7 @@ class ResamplingBinnedModel(ResamplingModel):
         # is faster with the current implementation of `sc.bin`.
         edges = self.edges[-1]
         dim = edges.dims[-1]
-        if dim in array.data.bins.data.coords:
+        if dim in array.data.bins.coords:
             # Must specify bounds for final dim despite handling by `histogram`
             # below: If coord is ragged binning would throw otherwise.
             bounds = sc.concatenate(edges[dim, 0], edges[dim, -1], dim)