+++ variable/include/scipp/variable/comparison.h
@@ -11,9 +11,11 @@ namespace scipp::variable {
 
 /// Functions to provide numpy-like element-wise comparison
 /// between two Variables.
+enum class NanComparisons { Equal, NotEqual };
 
-SCIPP_VARIABLE_EXPORT Variable is_approx(const VariableConstView &a,
-                                         const VariableConstView &b,
-                                         const VariableConstView &t);
+SCIPP_VARIABLE_EXPORT Variable
+isclose(const VariableConstView &a, const VariableConstView &b,
+        const VariableConstView rtol, const VariableConstView atol,
+        const NanComparisons equal_nans = NanComparisons::NotEqual);
 
 } // namespace scipp::variable