+++ variable/test/operations_test.cpp
@@ -375,10 +375,11 @@ TEST(Variable, concatenate_fail) {
   auto a = makeVariable<double>(Dimensions(dims), Values{1.0});
   auto b = makeVariable<double>(Dimensions(dims), Values{2.0});
   auto c = makeVariable<float>(Dimensions(dims), Values{2.0});
-  EXPECT_THROW_MSG(concatenate(a, c, Dim::X), std::runtime_error,
-                   "Cannot concatenate Variables: Data types do not match.");
+  EXPECT_THROW_MSG_DISCARD(
+      concatenate(a, c, Dim::X), std::runtime_error,
+      "Cannot concatenate Variables: Data types do not match.");
   auto aa = concatenate(a, a, Dim::X);
-  EXPECT_THROW_MSG(
+  EXPECT_THROW_MSG_DISCARD(
       concatenate(a, aa, Dim::Y), std::runtime_error,
       "Cannot concatenate Variables: Dimension extents do not match.");
 }
@@ -389,8 +390,8 @@ TEST(Variable, concatenate_unit_fail) {
   auto b(a);
   EXPECT_NO_THROW(concatenate(a, b, Dim::X));
   a.setUnit(units::m);
-  EXPECT_THROW_MSG(concatenate(a, b, Dim::X), std::runtime_error,
-                   "Cannot concatenate Variables: Units do not match.");
+  EXPECT_THROW_MSG_DISCARD(concatenate(a, b, Dim::X), std::runtime_error,
+                           "Cannot concatenate Variables: Units do not match.");
   b.setUnit(units::m);
   EXPECT_NO_THROW(concatenate(a, b, Dim::X));
 }